Move TTS functions out
